Demo constant time
Try to make this happen: use same cpu cycles in each function call on
q_insert, but fail at traces/trace-17-complexity.cmd test. Because of
not copying in element_t->value with s and instead using other
variables. Maximum length of input string on qtest interpreter is 4092
on keyboard key in. So suppose max size is 4096, and try to perform
copying 4096 strings in each function call on q_insert.

Change-Id: I569958e64d1b52a709bdacbfd9f658196a726caa